refactor(headless-styles): update imports for Position type in Tooltip code
